api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.17.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.27.0
yt-dlp==2025.1.26
supabase
youtube-transcript-api
